Tidy up file structure
