+++ src/dodal/devices/insertion_device/apple2_controller.py
@@ -1,25 +1,31 @@
 import abc
+import asyncio
 from dataclasses import dataclass
 from math import isclose
 from typing import Generic, Protocol, TypeVar
 
+from bluesky.protocols import Movable
 from ophyd_async.core import (
+    AsyncStatus,
     Reference,
     StandardReadable,
     StandardReadableFormat,
     derived_signal_rw,
     soft_signal_r_and_setter,
     soft_signal_rw,
+    wait_for_value,
 )
 
-from dodal.devices.insertion_device.apple2_undulator import Apple2, Apple2Val
+from dodal.devices.insertion_device.apple2_undulator_base import undulator_check_move
+from dodal.devices.insertion_device.apple2_undulator_gap import UndulatorGap
 from dodal.devices.insertion_device.apple2_undulator_phase_axes import (
+    Apple2LockedPhasesVal,
     Apple2PhasesVal,
     PhaseAxesType,
     UndulatorPhaseAxes,
 )
 from dodal.devices.insertion_device.energy_motor_lookup import EnergyMotorLookup
-from dodal.devices.insertion_device.enum import Pol
+from dodal.devices.insertion_device.enum import Pol, UndulatorGateStatus
 from dodal.log import LOGGER
 
 T = TypeVar("T")
@@ -27,6 +33,59 @@ MAXIMUM_MOVE_TIME = 550  # There is no useful movements take longer than this.
 ROW_PHASE_MOTOR_TOLERANCE = 0.004
 MAXIMUM_ROW_PHASE_MOTOR_POSITION = 24.0
 MAXIMUM_GAP_MOTOR_POSITION = 100
+
+
+@dataclass
+class Apple2Val:
+    gap: float
+    phase: Apple2LockedPhasesVal | Apple2PhasesVal
+
+    def extract_phase_val(self):
+        return self.phase
+
+
+class Apple2(StandardReadable, Movable[Apple2Val], Generic[PhaseAxesType]):
+    """Device representing the combined motor controls for an Apple2 undulator.
+
+    Attributes:
+        gap_ref (Reference[UndulatorGap]): The undulator gap motor device.
+        phase_ref (Reference[UndulatorPhaseAxes]): The undulator phase axes device,
+            consisting offour phase motors.
+
+    Args:
+        id_gap (UndulatorGap): An UndulatorGap device.
+        id_phase (UndulatorPhaseAxes): An UndulatorPhaseAxes device.
+        name (str, optional): Name of the device.
+    """
+
+    def __init__(self, id_gap: UndulatorGap, id_phase: PhaseAxesType, name=""):
+        with self.add_children_as_readables():
+            self.gap_ref = Reference(id_gap)
+            self.phase_ref = Reference(id_phase)
+        super().__init__(name=name)
+
+    @AsyncStatus.wrap
+    async def set(self, id_motor_values: Apple2Val) -> None:
+        """Check ID is in a movable state and set all the demand value before moving
+        them all at the same time.
+        """
+        gap = self.gap_ref()
+        phase = self.phase_ref()
+        # Only need to check gap as the phase motors share both status and gate with gap.
+        await undulator_check_move(gap.status, gap.gate)
+        await asyncio.gather(
+            phase.set_demand_positions(value=id_motor_values.extract_phase_val()),
+            gap.set_demand_positions(value=float(id_motor_values.gap)),
+        )
+        timeout = max(await asyncio.gather(gap.get_timeout(), phase.get_timeout()))
+        LOGGER.info(
+            f"Moving {self.name} apple2 motors to {id_motor_values}, timeout = {timeout}"
+        )
+        await asyncio.gather(
+            gap.set_move.set(value=1, timeout=timeout),
+            phase.set_move.set(value=1, timeout=timeout),
+        )
+        await wait_for_value(gap.gate, UndulatorGateStatus.CLOSE, timeout=timeout)
 
 
 class EnergyMotorConvertor(Protocol):
